Split rust grpc bindings into dedicated crate
When building gRPC bindings for Rust, we may have to compile additional
dependencies.

By splitting the gRPC bindings into their own crate, these dependencies
may be cached more effectively, especially in Docker.

Furthermore, futures-mpsc-lossy and tower-router have been moved into
the `proxy` directory.
