+++ proxy/grpc/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "conduit-grpc"
+version = "0.2.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+publish = false
+
+[dependencies]
+convert = { path = "../convert" }
+
+bytes = "0.4"
+futures = "0.1"
+h2 = "0.1"
+http = "0.1"
+prost = "0.3.0"
+prost-derive = "0.3.0"
+prost-types = "0.3.0"
+
+tower-grpc = { git = "https://github.com/tower-rs/tower-grpc" }
+tower-h2   = { git = "https://github.com/tower-rs/tower-h2" }
+
+[build-dependencies]
+tower-grpc-build = { git = "https://github.com/tower-rs/tower-grpc" }
