+++ proxy/src/telemetry/tap/match_.rs
@@ -6,8 +6,8 @@ use http;
 use ipnet::{Contains, Ipv4Net, Ipv6Net};
 
 use super::Event;
-use control::pb::common::ip_address;
-use control::pb::tap::observe_request;
+use conduit_grpc::common::ip_address;
+use conduit_grpc::tap::observe_request;
 use convert::*;
 use ctx;
 
@@ -136,7 +136,7 @@ impl<'a> TryFrom<&'a observe_request::match_::Match> for Match {
 
     #[allow(unconditional_recursion)]
     fn try_from(m: &observe_request::match_::Match) -> Result<Self, Self::Err> {
-        use control::pb::proxy::tap::observe_request::match_;
+        use conduit_grpc::tap::observe_request::match_;
 
         let match_ = match *m {
             match_::Match::All(ref seq) => Match::All(Self::from_seq(seq)?),
@@ -177,7 +177,7 @@ impl<'a> TryFrom<&'a observe_request::match_::Tcp> for TcpMatch {
     type Err = InvalidMatch;
 
     fn try_from(m: &observe_request::match_::Tcp) -> Result<Self, InvalidMatch> {
-        use control::pb::proxy::tap::observe_request::match_::tcp;
+        use conduit_grpc::tap::observe_request::match_::tcp;
 
         let m = match m.match_.as_ref() {
             None => return Err(InvalidMatch::Empty),
@@ -280,7 +280,7 @@ impl HttpMatch {
         string_match: &observe_request::match_::http::string_match::Match,
         value: &str,
     ) -> bool {
-        use control::pb::proxy::tap::observe_request::match_::http::string_match::Match::*;
+        use conduit_grpc::tap::observe_request::match_::http::string_match::Match::*;
 
         match *string_match {
             Exact(ref exact) => value == exact,
@@ -292,7 +292,7 @@ impl HttpMatch {
 impl<'a> TryFrom<&'a observe_request::match_::Http> for HttpMatch {
     type Err = InvalidMatch;
     fn try_from(m: &'a observe_request::match_::Http) -> Result<Self, InvalidMatch> {
-        use control::pb::proxy::tap::observe_request::match_::http::Match as Pb;
+        use conduit_grpc::tap::observe_request::match_::http::Match as Pb;
 
         m.match_
             .as_ref()
@@ -338,8 +338,8 @@ mod tests {
     use quickcheck::*;
 
     use super::*;
-    use control::pb::common::*;
-    use control::pb::proxy::tap::*;
+    use conduit_grpc::common::*;
+    use conduit_grpc::tap::*;
 
     impl Arbitrary for ObserveRequest {
         fn arbitrary<G: Gen>(g: &mut G) -> Self {
